requirements.txt
Django==4.2.4
django_import_export==3.2.0
django_mptt==0.14.0
pandas==2.0.3
